pypiorg/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.5.3
Flask-Login==0.5.0
Flask-WTF==0.14.3
email-validator==1.1.1
passlib==1.7.2
